supportSystem/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework_gis==1.0
psycopg2==2.9.4
